requirements.txt
captum==0.8.0
gradio==6.12.0
groq==1.1.2
huggingface_hub==1.7.2
joblib==1.5.3
numpy
plotly==6.6.0
python-dotenv==1.2.2
rank_bm25==0.2.2
sentence_transformers==5.3.0
torch==2.5.1
transformers==5.3.0
